workers and cashier added

+++ SuperShop/SuperShopDatabase/SuperShopDatabase/dao/inter/IBarcodeDAO.cs
@@ -10,7 +10,7 @@ namespace SuperShopDatabase.Dao.Inter
     public interface IBarcodeDAO
     {
         List<Barcode> GetAll ();
-        Barcode GetBarcodeById (int id);
+        Barcode GetBarcodeById (long id);
         Barcode AddBarcode (Barcode barcode);
         Barcode UpdateBarcode (Barcode barcode);
         Barcode RemoveBarcode (Barcode barcode);
